LPS-68122 Cannot invalidate sessions in valueUnbound
HttpSessionBindingListener.valueUnbound is not appropiate for children
session invalidation since the parent session is already invalidated by
then.

This pull makes ProxyServlet register a HttpSessionListener in the
ServletContext that is then used to invalidate related children sessions
when appropiate.

This will only work the registration of ProxyServlet is done at a time
in which the ServletContext allows to register listeners.